requirements.txt
Flask
assemblyai
requests
elevenlabs
vertexai
google-generativeai
markdown
python-dotenv